+++ deepchem/models/gan.py
